requirements_backup_GeoSPSR_uv.txt
absl-py==2.3.1
asttokens==3.0.0
chex==0.1.89
colorama==0.4.6
comm==0.2.2
contourpy==1.3.2
cycler==0.12.1
decorator==5.2.1
etils==1.13.0
exceptiongroup==1.3.0
executing==2.2.0
flax==0.10.7
fonttools==4.59.0
fsspec==2025.7.0
gpytoolbox==0.3.6
humanize==4.12.3
imageio==2.37.0
importlib-resources==6.5.2
ipydatawidgets==4.3.5
ipython==8.37.0
ipywidgets==8.1.7
jax==0.6.2
jaxlib==0.6.2
jedi==0.19.2
jupyterlab-widgets==3.0.15
kiwisolver==1.4.8
lazy-loader==0.4
libigl==2.6.1
markdown-it-py==3.0.0
matplotlib==3.10.3
matplotlib-inline==0.1.7
mdurl==0.1.2
meshplot @ git+https://github.com/skoch9/meshplot.git@725e4a7926a5f10888f0edd1762fecf9db751c56
ml-dtypes==0.5.1
msgpack==1.1.1
nest-asyncio==1.6.0
networkx==3.4.2
numpy==2.2.6
opt-einsum==3.4.0
optax==0.2.5
orbax-checkpoint==0.11.19
packaging==25.0
parso==0.8.4
pillow==11.3.0
pip==22.2.1
prompt-toolkit==3.0.51
protobuf==6.31.1
pure-eval==0.2.3
pygments==2.19.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pythreejs==2.4.2
pyyaml==6.0.2
rich==14.0.0
scikit-image==0.25.2
scipy==1.15.0
scs==3.2.7.post2
setuptools==63.2.0
simplejson==3.20.1
six==1.17.0
stack-data==0.6.3
tensorstore==0.1.76
tifffile==2025.5.10
toolz==1.0.0
tqdm==4.67.1
traitlets==5.14.3
traittypes==0.2.1
treescope==0.1.9
trimesh==4.7.1
typing-extensions==4.14.1
wcwidth==0.2.13
widgetsnbextension==4.0.14
zipp==3.23.0
